speeds are floats

+++ fw/src/control/speed.h
@@ -3,12 +3,12 @@
 #include <stdint.h>
 
 // The measured motor speeds via encoder period in microseconds.
-extern int16_t speed_measured_left;
-extern int16_t speed_measured_right;
+extern float speed_measured_left;
+extern float speed_measured_right;
 
 // The speeds setpoints via encoder period in microsecond
-extern int16_t speed_setpoint_left;
-extern int16_t speed_setpoint_right;
+extern float speed_setpoint_left;
+extern float speed_setpoint_right;
 
 void speed_init();
 void speed_update();
@@ -16,5 +16,5 @@ void speed_update();
 void speed_enable();
 void speed_disable();
 
-void speed_set_left(int16_t setpoint);
-void speed_set_right(int16_t setpoint);
+void speed_set_left(float setpoint);
+void speed_set_right(float setpoint);
